Move proxy reference to win32/win64 subfolders
To keep all FFmpeg dll dependencies in the root directory, the x64 dlls would have to be renamed (x64 infix like done with the other library dependencies), which means it would need to be changed in the sources and compiled from there. Just renaming the dll files and creating new lib files with the changed names does not work, because the dll files themselves also contain references to each other by their original filenames.