Move output_buffer from proxy to user
before this change, the output_buffer has been allocated by the proxy, and handed up to the user. While this worked, it made handling in C# code more complicated and required copying of the data into the managed domain. Now the buffer can be created in C# and handed directly to the proxy.