Add support for files which cannot be seeked by PTS
Files, for which FFmpeg does not provide a length and packet/frame PTS, cannot be seeked by PTS. This is either because it is not supported by the file format (e.g. SHN) or not implemented in FFmpeg (e.g. SHN+seektable extension). To work around this issue, the file is sequentially decoded into a wave proxy that is then used by AudioAlign in place of the non-seekable file. With this change, the proxy also returns the untouched AV_NOPTS_VALUE instead of assuming it is a valid timestamp and converting it to sample time.