Refactor sqlite-net for Windows x86 and x64 support
All bindings in sqlite-net go statically to sqlite3.dll, which can be an x86 or x64 dll depending on the build target platform. Because of this design, an application cannot ship with dlls for both architectures and decide at runtime which one to use, except when using workarounds like SetDllDirectory (which can screw with other libraries or components that also bind dlls) or renaming/moving dlls from a dll repository. The changes introduced here enable the dynamic usage of the correct dll for the current architecture, but with some overhead (one level of indirection when calling functions).
As an alternative, the deployment could be changed to only deploy the dlls of the current architecture, in which case these changes could be reverted to the initial state.