Fix warping of the first interval
The start of the first warping interval wasn't added to the time warps, leading to too much warping / a too long first warping interval. it resulted in warping from the start of the stream to the end of the warp interval, instead of from the start of the warp interval to the end of the warp interval.