Fix a nasty bug in resampling buffer management
The buffer in the resampling stream got resized without taking care about the content of the buffer, leading to lost / zeroed out data that has not been consumed yet. This lead to resulting zero samples or near-zero samples after resampling, screwing up the whole stream. By moving the buffer resizing code to a place where it is guaranteed to be empty (just before being refilled), this issue has been solved.