Keep resampler instance through time warps instead of instantiating a new one for each time warp section
Creating a new instance of the resampler through a new resampler stream and and accompanying crop stream can happen a lot with short time warp sections. It is unknown if it caused a noticeable overhead but it feels better if it does not happen, also keeping the same resampler can be useful when implementing interpolation between different sample rates.