Change hash from struct to uint for a huge performance gain
The wrapping struct turned out to introduce a lot of overhead for accessing and comparing the hash value (probably due to the function call indirections). Changing it to a simple uint type made it possible to use the Zipper intersection algorithm for faster hash collision detection in the match detection. This resulted in a computation speedup of ~300% for the match detection call in the fingerprint store.