Add Chromaprint chroma mapping mode
This commit adds a selection of two FFT bin to chroma bin mapping modes. The first one is the mode as described in the referenced paper, and the second is the mode as used in the original Chromaprint algorithm. When using Chromaprint mode, the resulting fingerprints are validated to be exactly the same as calculated by the original Chromaprint implementation, thus they are compatible. The paper mode seems to increase the matching rate.