Improve Hann windows again
Normal Hann window calculation has been reverted to previous version, which was correct in calculating a symmetric window with zero samples. This is similar to Matlab's hann() function, in contrast to the hanning() function that omits the zeros which caused this confusion. A new periodic Hann window has been added, useful for FFT where even lengths are required, to still fulfil the COLA property.