Fix STFT with increased frequency resolution
The zero-padded FFT input was overwritten by the FFT output because the FFP was performed in-place. In the next round, only the window part was overwritten, the remaining samples contained FFT results instead of zero-padding samples, resulting in invalid FFT results. To solve it, the array must either be reinitialized with zeros before filling with audio samples, or a separate buffer for FFT output must be used. I chose the separate buffer, as it requires a bit additional memory but save the time that it takes clearing the buffer before every read.