Release references to mappings passed in from outside
Fixes a bug where disposed time warp streams would still have a reference to mappings passed in from the outside, listen to their change events, and work on streams which have already been disposed and thus cannot be used any longer. But working on disposed streams was just the collateral, the actual issue was the reference to an outside mappings list.