feat: direct audio proxy file writing (#25)
When generating an audio proxy file for an FFmpeg stream, it was first fully decoded and stored in memory before it was written to a file. This was an inefficient waste of memory resources, and proxy files were limited to a size of 2 GiB due to a limitation of .NET's `MemoryStream`. With this change, the files are written directly. They are now limited to 4 GiB due to a limitation of NAudio's `WaveFileWriter`.